Convert public API numeric types to Integer (resolve #43) (#47)
* Change MediaInfo length type from double to int

* Change AdObject and AdBreakObject types to 'int' for position, startTime, and length

* Change QoE Details types to 'int'

* Change playhead type to Int

* Change chapter details type to integer in MediaAnalyticsProvider

* Change preroll wait time type to 'int'

* Change types to int in Kotlin test app

* Change preroll wait time type to int in tests

* Correct configuration key names in MediaRealTimeSession logs

* Change datatypes to 'int' in integration test cases

* Delete dummy FunctionalTest file

+++ code/edgemedia/src/androidTest/java/com/adobe/marketing/mobile/util/EdgeEventHelper.kt
@@ -140,7 +140,7 @@ internal class EdgeEventHelper {
             val details: MutableMap<String, Any> = mutableMapOf(
                 "name" to (info?.get("media.id") ?: ""),
                 "friendlyName" to (info?.get("media.name") ?: ""),
-                "length" to ((info?.get("media.length") as? Double) ?: -1).toInt(),
+                "length" to ((info?.get("media.length") as? Int) ?: -1),
                 "streamType" to (info?.get("media.type") ?: ""),
                 "contentType" to (info?.get("media.streamtype") ?: ""),
                 "hasResume" to (info?.get("media.resumed") ?: "")
@@ -203,8 +203,8 @@ internal class EdgeEventHelper {
             val details: MutableMap<String, Any> = mutableMapOf(
                 "name" to (info?.get("ad.id") ?: ""),
                 "friendlyName" to (info?.get("ad.name") ?: ""),
-                "podPosition" to ((info?.get("ad.position") as? Long) ?: -1).toInt(),
-                "length" to ((info?.get("ad.length") as? Double) ?: -1).toInt()
+                "podPosition" to ((info?.get("ad.position") as? Int) ?: -1),
+                "length" to ((info?.get("ad.length") as? Int) ?: -1)
             )
 
             metadata?.forEach { (key, value) ->
@@ -228,17 +228,17 @@ internal class EdgeEventHelper {
         private fun getAdvertisingPodDetails(info: Map<String, Any>?): Map<String, Any> {
             return mapOf(
                 "friendlyName" to (info?.get("adbreak.name") ?: ""),
-                "index" to ((info?.get("adbreak.position") as? Long) ?: -1).toInt(),
-                "offset" to ((info?.get("adbreak.starttime") as? Double) ?: -1).toInt()
+                "index" to ((info?.get("adbreak.position") as? Int) ?: -1),
+                "offset" to ((info?.get("adbreak.starttime") as? Int) ?: -1)
             )
         }
 
         private fun getChapterDetails(info: Map<String, Any>?): Map<String, Any> {
             return mapOf(
                 "friendlyName" to (info?.get("chapter.name") ?: ""),
-                "index" to ((info?.get("chapter.position") as? Long) ?: -1).toInt(),
-                "offset" to ((info?.get("chapter.starttime") as? Double) ?: -1).toInt(),
-                "length" to ((info?.get("chapter.length") as? Double) ?: -1).toInt()
+                "index" to ((info?.get("chapter.position") as? Int) ?: -1),
+                "offset" to ((info?.get("chapter.starttime") as? Int) ?: -1),
+                "length" to ((info?.get("chapter.length") as? Int) ?: -1)
             )
         }
 
@@ -259,10 +259,10 @@ internal class EdgeEventHelper {
 
         private fun getQoeDetails(info: Map<String, Any>?): Map<String, Any> {
             return mapOf(
-                "bitrate" to ((info?.get("qoe.bitrate") as? Double) ?: -1).toInt(),
-                "droppedFrames" to ((info?.get("qoe.droppedframes") as? Double) ?: -1).toInt(),
-                "framesPerSecond" to ((info?.get("qoe.fps") as? Double) ?: -1).toInt(),
-                "timeToStart" to ((info?.get("qoe.startuptime") as? Double) ?: -1).toInt()
+                "bitrate" to ((info?.get("qoe.bitrate") as? Int) ?: -1),
+                "droppedFrames" to ((info?.get("qoe.droppedframes") as? Int) ?: -1),
+                "framesPerSecond" to ((info?.get("qoe.fps") as? Int) ?: -1),
+                "timeToStart" to ((info?.get("qoe.startuptime") as? Int) ?: -1)
             )
         }
     }
