+++ code/edgemedia/src/test/java/com/adobe/marketing/mobile/edge/media/internal/EdgeEventHelper.kt
@@ -23,7 +23,7 @@ internal class EdgeEventHelper {
     companion object {
         fun generateEdgeEvent(
             eventType: XDMMediaEventType,
-            playhead: Long,
+            playhead: Int,
             timestamp: Long,
             backendSessionId: String? = null,
             info: Map<String, Any>? = null,
@@ -65,7 +65,7 @@ internal class EdgeEventHelper {
 
         private fun generateMediaCollection(
             eventType: XDMMediaEventType,
-            playhead: Long,
+            playhead: Int,
             backendSessionId: String?,
             info: Map<String, Any>?,
             metadata: Map<String, Any>?,
@@ -132,7 +132,7 @@ internal class EdgeEventHelper {
             val details: MutableMap<String, Any> = mutableMapOf(
                 "name" to (info?.get("media.id") ?: ""),
                 "friendlyName" to (info?.get("media.name") ?: ""),
-                "length" to ((info?.get("media.length") as? Double) ?: -1).toLong(),
+                "length" to ((info?.get("media.length") as? Int) ?: -1),
                 "streamType" to (info?.get("media.type") ?: ""),
                 "contentType" to (info?.get("media.streamtype") ?: ""),
                 "hasResume" to (info?.get("media.resumed") ?: "")
@@ -195,8 +195,8 @@ internal class EdgeEventHelper {
             val details: MutableMap<String, Any> = mutableMapOf(
                 "name" to (info?.get("ad.id") ?: ""),
                 "friendlyName" to (info?.get("ad.name") ?: ""),
-                "podPosition" to ((info?.get("ad.position") as? Long) ?: -1).toLong(),
-                "length" to ((info?.get("ad.length") as? Double) ?: -1).toLong()
+                "podPosition" to ((info?.get("ad.position") as? Int) ?: -1),
+                "length" to ((info?.get("ad.length") as? Int) ?: -1)
             )
 
             metadata?.forEach { (key, value) ->
@@ -220,17 +220,17 @@ internal class EdgeEventHelper {
         private fun getAdvertisingPodDetails(info: Map<String, Any>?): Map<String, Any> {
             return mapOf(
                 "friendlyName" to (info?.get("adbreak.name") ?: ""),
-                "index" to ((info?.get("adbreak.position") as? Long) ?: -1).toLong(),
-                "offset" to ((info?.get("adbreak.starttime") as? Double) ?: -1).toLong()
+                "index" to ((info?.get("adbreak.position") as? Int) ?: -1),
+                "offset" to ((info?.get("adbreak.starttime") as? Int) ?: -1)
             )
         }
 
         private fun getChapterDetails(info: Map<String, Any>?): Map<String, Any> {
             return mapOf(
                 "friendlyName" to (info?.get("chapter.name") ?: ""),
-                "index" to ((info?.get("chapter.position") as? Long) ?: -1).toLong(),
-                "offset" to ((info?.get("chapter.starttime") as? Double) ?: -1).toLong(),
-                "length" to ((info?.get("chapter.length") as? Double) ?: -1).toLong()
+                "index" to ((info?.get("chapter.position") as? Int) ?: -1),
+                "offset" to ((info?.get("chapter.starttime") as? Int) ?: -1),
+                "length" to ((info?.get("chapter.length") as? Int) ?: -1)
             )
         }
 
@@ -251,10 +251,10 @@ internal class EdgeEventHelper {
 
         private fun getQoeDetails(info: Map<String, Any>?): Map<String, Any> {
             return mapOf(
-                "bitrate" to ((info?.get("qoe.bitrate") as? Double) ?: -1).toLong(),
-                "droppedFrames" to ((info?.get("qoe.droppedframes") as? Double) ?: -1).toLong(),
-                "framesPerSecond" to ((info?.get("qoe.fps") as? Double) ?: -1).toLong(),
-                "timeToStart" to ((info?.get("qoe.startuptime") as? Double) ?: -1).toLong()
+                "bitrate" to ((info?.get("qoe.bitrate") as? Int) ?: -1),
+                "droppedFrames" to ((info?.get("qoe.droppedframes") as? Int) ?: -1),
+                "framesPerSecond" to ((info?.get("qoe.fps") as? Int) ?: -1),
+                "timeToStart" to ((info?.get("qoe.startuptime") as? Int) ?: -1)
             )
         }
     }
